Move thermostat functions to module
